Improve dice complexity
